🚚 Move: Rename Kanji entity to File and adjust DTOs/enums
- Introduce `entity.File` (extends BaseEntity) with `member` relation
- Move/rename `TaskStatus` to `domain.kanji.enumerated`
- Add `FileUploadResponse` (taskId/fileName/status)
- Update `KanjiVocabularyListResponse.from(...)` to accept `File`

ref: #60
